requirements.txt
numpy~=2.0
pandas~=2.2
python-dotenv~=1.0
supabase~=2.4
httpx~=0.28
PySimpleGUI~=5.0
customtkinter~=5.2.2
packaging
Pillow
psycopg2-binary
